No Spring

+++ src/main/java/dealerstat/service/AuthenticationService.java
@@ -43,26 +43,28 @@ public class AuthenticationService {
 
     public ResponseEntity<?> registerUser(@RequestBody @Valid MyUserDto myUserDto) {
         String email = (String) redisService.getToken(myUserDto.getEmail());
-        MyUser myUser = myUserService.findMyUserByEmail(myUserDto.getEmail());
+        MyUser myUser = myUserService.findMyUserByEmail(myUserDto.getEmail().toLowerCase(Locale.ROOT));
 
         if (myUser == null && email == null) {
             return createMyUserForRegistration(myUserDto, null, null, 0.0);
         } else if (email != null) {
             return ResponseEntity.status(404).body("Check your email you have already registered");
-        } else if (myUser != null && myUser.getPassword() != null) {
+        } else if (myUser.getPassword() != null) {
             return ResponseEntity.status(404).body("User with email " + myUser.getEmail() + " already exist");
         } else {
             return createMyUserForRegistration(myUserDto, myUser.getId(), myUser.getCreatedAt(), myUser.getRating());
         }
     }
 
-    private ResponseEntity createMyUserForRegistration(MyUserDto myUserDto, Long id, Date createdAt,
+    private ResponseEntity<?> createMyUserForRegistration(MyUserDto myUserDto, Long id, Date createdAt,
                                                        Double rating) {
         MyUser myUser = new MyUser();
+
         if (id != null) {
             myUser.setId(id);
             myUser.setCreatedAt(createdAt);
         }
+
         myUser.setFirstName(myUserDto.getFirstName());
         myUser.setLastName(myUserDto.getLastName());
         myUser.setPassword(passwordEncoder.encode(myUserDto.getPassword()));
@@ -76,6 +78,7 @@ public class AuthenticationService {
         messageSender(myUserDto.getEmail(), token, "confirm");
         redisService.putToken(token, myUser);
         redisService.putToken(myUserDto.getEmail(), "CHECK");
+
         return ResponseEntity.ok("To complete the registration, check your email, please.");
     }
 
@@ -90,7 +93,6 @@ public class AuthenticationService {
             } else if (!user.isApproved()) {
                 return ResponseEntity.status(403).body("Admin hasn't yet confirmed your request");
             } else {
-
                 String token = jwtTokenProvider.createToken(email, user.getId(), user.getRoles());
 
                 Map<Object, Object> response = new HashMap<>();
@@ -107,6 +109,7 @@ public class AuthenticationService {
 
     public ResponseEntity<?> confirmAccount(String token) {
         MyUser myUser = (MyUser) redisService.getToken(token);
+
         if (myUser != null) {
             myUserService.save(myUser);
             redisService.deleteToken(myUser.getEmail());
@@ -118,11 +121,12 @@ public class AuthenticationService {
     }
 
     public ResponseEntity<?> sendMessageForSetPassword(String email) {
-        MyUser myUser = myUserService.findMyUserByEmail(email);
+        MyUser myUser = myUserService.findMyUserByEmail(email.toLowerCase(Locale.ROOT));
+
         if (myUser != null) {
             String token = UUID.randomUUID().toString();
 
-            messageSender(email, token, "set_password");
+            messageSender(email, token, "reset");
 
             redisService.putToken(token, email);
 
@@ -133,22 +137,28 @@ public class AuthenticationService {
     }
 
     public ResponseEntity<?> setPassword(String password, String token) {
-        String email = (String) redisService.getToken(token);
-        if (email != null) {
-            MyUser myUser = myUserService.findMyUserByEmail(email);
-            myUser.setPassword(passwordEncoder.encode(password));
-            myUserService.save(myUser);
-            redisService.deleteToken(token);
-            return ResponseEntity.ok("Your password has been successfully changed!");
+        if (password.length() >= 8) {
+            String email = (String) redisService.getToken(token);
+            if (email != null && myUserService.findMyUserByEmail(email) != null) {
+                MyUser myUser = myUserService.findMyUserByEmail(email);
+                myUser.setPassword(passwordEncoder.encode(password));
+                myUserService.save(myUser);
+                redisService.deleteToken(token);
+                return ResponseEntity.ok("Your password has been successfully changed!");
+            } else {
+                return ResponseEntity.status(404).body("The token's lifetime has expired");
+            }
         } else {
-            return ResponseEntity.status(404).body("The token's lifetime has expired");
+            return ResponseEntity.status(404).body("Password should be more than 8");
         }
+
     }
 
     public ResponseEntity<?> checkCode(String token) {
         String check = (String) redisService.getToken(token);
+
         if (check != null) {
-            return ResponseEntity.ok(check);
+            return ResponseEntity.ok("The token's lifetime has not expired");
         } else {
             return ResponseEntity.status(404).body("The token's lifetime has expired");
         }
